+++ pkg/storage/localstore/export_test.go
@@ -75,7 +75,7 @@ func TestExportImport(t *testing.T) {
 		}
 		got := ch.Data()
 		if !bytes.Equal(got, want) {
-			t.Fatalf("chunk %s: got data %s, want %x", addr, got, want)
+			t.Fatalf("chunk %s: got data %x, want %x", addr, got, want)
 		}
 	}
 }
